chore: pulled out fecth data from App.svetle
