+++ examples/headless-example/src/services/ballerine-backoffice.service.ts
@@ -1,0 +1,39 @@
+import { fetchJson } from '@/utils';
+
+export type WorkServiceEndpoints = {
+  base: string;
+};
+export class BallerineBackOfficeService {
+  constructor(private readonly baseUrl: string = 'http://localhost:3000/api/external') {}
+
+  fetchEndUser = async (id: string) => fetchJson(`${this.baseUrl}/end-users/${id}`);
+
+  fetchWorkflow = async (id: string) => fetchJson(`${this.baseUrl}/workflows/${id}`);
+  fetchWorkflows = async () =>
+    fetchJson<
+      Array<{
+        workflowDefinition: {
+          id: string;
+          name: string;
+        };
+        workflowRuntimeData: {
+          id: string;
+          status: string;
+        };
+      }>
+    >(`${this.baseUrl}/workflows`);
+  fetchIntent = async () =>
+    fetchJson<Array<Record<string, unknown>>>(`${this.baseUrl}/workflows/intent`, {
+      method: 'POST',
+      body: { intentName: 'kyc' },
+    });
+
+  fetchSignUp = async ({ firstName, lastName }: { firstName: string; lastName: string }) =>
+    fetchJson(`${this.baseUrl}/end-users`, {
+      method: 'POST',
+      body: {
+        firstName,
+        lastName,
+      },
+    });
+}
