--- examples/headless-example/src/services/workflow.service.ts
@@ -1,6 +1,0 @@
-export type WorkServiceEndpoints = {
-  base: string;
-};
-export class WorkflowService {
-  constructor(endpoints: WorkServiceEndpoints) {}
-}
